fixes and changes
