+++ farmodo/constants.py
@@ -1,25 +1,1 @@
 # constants
-
-
-JOBID ='id'
-SCENE ='scenePath'
-JOBSTART='start'
-JOBEND = 'end'
-JOBSTEP = 'step'
-CAMERA = 'camera'
-INITIAL_PASS = 'initialPass'
-PASS = 'passes'
-PASSGRP = 'passGroup'
-IMAGEFILE ='imageFile'
-IMAGEFORMAT = 'imageFormat'
-JOBSTATUS= 'status'
-PRIORITY = 'priority'
-LIMIT ='timeLimit'
-TIMEIN = 'timeSubmitted'
-USER = 'username'
-IPADDRESS = 'ipAddress'
-APP = 'application'
-FRAMEWIDTH ='frameWidth'
-FRAMEHEIGHT = 'frameHeight'
-SAMPLES ='samples'
-IMPORTANCE ='importance'
