+++ farmodo/renderModo.py
@@ -6,13 +6,38 @@ import socket
 from datetime import date, datetime
 import platform
 import modo
-import constants
 global renderJobId
 
-jobsPending = True
 
+JOBID ='id'
+SCENE ='scenePath'
+JOBSTART='start'
+JOBEND = 'end'
+JOBSTEP = 'step'
+CAMERA = 'camera'
+INITIAL_PASS = 'initialPass'
+PASSLIST = 'passes'
+PASSGRP = 'passGroup'
+IMAGEFILE ='imageFile'
+IMAGEFORMAT = 'imageFormat'
+JOBSTATUS= 'status'
+PRIORITY = 'priority'
+LIMIT ='timeLimit'
+TIMEIN = 'timeSubmitted'
+USER = 'username'
+IPADDRESS = 'ipAddress'
+APP = 'application'
+FRAMEWIDTH ='frameWidth'
+FRAMEHEIGHT = 'frameHeight'
+SAMPLES ='samples'
+IMPORTANCE ='importance'
+tagFileName = '/tmp/blt_lockFile'
+debug = 2
+
+jobsPending = True
+print 'render modo'
 def renderJob(job):    #render.animation this job
-    debug = "0"
+
     # usage @modoRenderSubmit path/to/scene start end step output/path format -pass "comma,sep,text" -camera default -pattern default
     if job:
         renderJobId = job[JOBID]
@@ -34,7 +59,7 @@ def renderJob(job):    #render.animation this job
     lx.out("step: %s " % job[JOBSTEP])
     lx.out("camera: %s " % job[CAMERA])
     lx.out("initialPass: %s " % job[INITIAL_PASS])
-    lx.out("passes: %s " % job[PASS])
+    lx.out("passes: %s " % job[PASSLIST])
     lx.out("passGroup: %s " % job[PASSGRP])
     lx.out("imageFile: %s " % job[IMAGEFILE])
     lx.out("imageFormat: %s " % job[IMAGEFORMAT])
@@ -59,7 +84,7 @@ def renderJob(job):    #render.animation this job
                 continue
         #intial pass setting - for multiple pass combos and multi node pass renders
         initialPassesPattern = ''
-        initialList = job[6]
+        initialList = job[INITIAL_PASS]
         if len(initialList):
             for eachInit in initialList.split(","):
                 eachInitHalf = eachInit.split(":")
@@ -80,22 +105,23 @@ def renderJob(job):    #render.animation this job
         if "<initialPass>" in outPat:
             print "initial pass in output pattern"
             newOutPat = outPat.replace('<initialPass>',initialPassesPattern)
-            if '__' in newOutPat:
-                newOutPat = newOutPat.replace('__','_')
+            newOutPat = newOutPat.replace('__','_')
+            newOutPat = newOutPat.replace('[','')
+            newOutPat = newOutPat.replace(']','')
             scene.items('polyRender')[0].channel('outPat').set(newOutPat)
             lx.out("Output pattern changed to :%s " % (scene.items('polyRender')[0].channel('outPat').get()))
 
 
 
-        lx.out("pass list specified %s" % job[7])
+        lx.out("pass list specified %s" % job[PASSLIST])
         #first switch everytning off
         n = lx.eval('query sceneservice actionItem.N ? ""')
         for x in range(n):  # enumerate ALL actionclips in scene
                 eachPass = lx.eval("query sceneservice actionItem.name ? %d" % x)
                 lx.eval("layer.enable {%s} off" % (eachPass))
                 lx.out("switched off %s" % (eachPass))
-        passDoList = job[7].split(",")
-        renderPassGroup = job[8]
+        passDoList = job[PASSLIST].split(':')[1].split(",")
+        renderPassGroup = job[PASSLIST].split(':')[0]
         if not renderPassGroup: #in case render pass group not given we can imply it from the pass
             for rpg in scene.renderPassGroups:
                 for eachPass in rpg.passes:
@@ -111,40 +137,40 @@ def renderJob(job):    #render.animation this job
         # # # # # #
 
         lx.eval("select.item Render")
-        lx.eval("item.channel first %s" % job[2])
-        lx.eval("item.channel last %s" % job[3])
-        lx.eval("item.channel step %s" % job[4])
+        lx.eval("item.channel first %s" % job[JOBSTART])
+        lx.eval("item.channel last %s" % job[JOBEND])
+        lx.eval("item.channel step %s" % job[JOBSTEP])
 
 
 
         # # # # # #
         #render animation
         # # # # # #
-        renderCmd = "render.animation filename:" + job[9] + " format:" + job[10]
+        renderCmd = "render.animation filename:" + job[SCENE] + " format:" + job[IMAGEFORMAT]
 
         lx.out(renderCmd)
-        if job[7] and (job[6] is not None):
-            renderCmd += " group:{" + renderPassGroup + "}"
+        # if job[7] and (job[6] is not None):
+        renderCmd += " group:{" + renderPassGroup + "}"
 
-        if(debug == "0"):
+        if(debug == 0):
             lx.out('Render Command: %s'% renderCmd)
             lx.eval(renderCmd)
             lx.eval ('!scene.close')
-        if(debug == "1"):
+        if(debug == 1):
             lx.out('Render Command: %s'% renderCmd)
-        if(debug == "2"):
+        if(debug == 2):
             lx.out('Saving scene only')
             lx.eval ('!scene.save')
 
-        lx.out('Render Finshed For: %s'% job[1])
-        markJobAsFinished(job[0],True)
+        lx.out('Render Finshed For: %s'% job[JOBID])
+        markJobAsFinished(job[JOBID],True)
         lx.eval('log.masterSave "/tmp/modoConsoleLog.txt"')
         return True
     except:
         print"******render Failed*******"
         print"Closing scene"
         lx.eval ('!scene.close')
-        markJobAsFinished(job[0],False)
+        markJobAsFinished(job[JOBID],False)
         return False
 
 
@@ -153,6 +179,7 @@ def renderJob(job):    #render.animation this job
 def jobAssignedToThisNode():
     #get job assigned to this node, there can only be one
     jobD={}
+    job=()
     conn = sqlite3.connect(lx.eval("user.value farmodo_pathToDB ?"))
     lx.out("Retrieving assigned job")
     username = socket.gethostname()
@@ -162,8 +189,9 @@ def jobAssignedToThisNode():
     cur.execute('SELECT * FROM renderJobs WHERE username = "%s" AND status = "%s" '  % (username,"rendering"))
     job=cur.fetchone()
     conn.close()
-    for i in range(len(job)):
-	    jobD[pragma[i][1]]=job[i]
+    if job:
+        for i in range(len(job)):
+    	    jobD[pragma[i][1]]=job[i]
     return jobD
 
 
@@ -188,21 +216,23 @@ def markJobAsFinished(renderJobId,success):
 
 
 def doIStillHaveAJob():
+    if (debug == 2):
+        exit()
     conn = sqlite3.connect(lx.eval("user.value farmodo_pathToDB ?"))
     print "Looking for another job"
     username = socket.gethostname()
     cur = conn.cursor()
-    cur.execute('SELECT id,scenePath FROM renderJobs WHERE status LIKE "pending" ORDER BY id ASC' ):
+    cur.execute('SELECT id,scenePath FROM renderJobs WHERE status LIKE "pending" ORDER BY id ASC' )
     job = cur.fetchone()
-        if job:
-            print("Job  %s found to render %s" %(job[0],job[1]))
-            #assign node to this job
-            cur.execute('UPDATE renderJobs SET status="rendering", username="%s" WHERE id=%d' % (username,job[0]))
-            print("Rendering assigned to:" + username)
-            conn.commit()
-            print"commit"
-            conn.close()
-            return True
+    if job:
+        print("Job  %s found to render %s" %(job[0],job[1]))
+        #assign node to this job
+        cur.execute('UPDATE renderJobs SET status="rendering", username="%s" WHERE id=%d' % (username,job[0]))
+        print("Rendering assigned to:" + username)
+        conn.commit()
+        print"commit"
+        conn.close()
+        return True
     conn.close()
     return False
 
@@ -215,8 +245,9 @@ if __name__ == "__main__":
         jobsPending = doIStillHaveAJob()
 
     try:
-        os.system('rm ' + blt_farm.tagFileName)
+        os.system('rm ' + tagFileName)
     except:
         print "Error trying to delete tmp file"
     #quit modo and nap
-    lx.eval('app.quit')
+    if(debug != 2):
+        lx.eval('app.quit')
